refactor to move evaluation out
