+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/index.ts
@@ -1,6 +1,7 @@
 export * from './agent-prompt';
 export * from './connections';
 export * from './from-ai';
+export * from './nodes';
 export * from './tools';
 export * from './trigger';
 export * from './workflow-similarity';
