+++ packages/@n8n/ai-workflow-builder.ee/evaluations/programmatic/evaluators/nodes.ts
@@ -1,0 +1,15 @@
+import type { INodeTypeDescription } from 'n8n-workflow';
+
+import type { SimpleWorkflow } from '@/types';
+import { validateNodes } from '@/validation/checks';
+import type { SingleEvaluatorResult } from '@/validation/types';
+
+import { calcSingleEvaluatorScore } from '../../utils/score';
+
+export function evaluateNodes(
+	workflow: SimpleWorkflow,
+	nodeTypes: INodeTypeDescription[],
+): SingleEvaluatorResult {
+	const violations = validateNodes(workflow, nodeTypes);
+	return { violations, score: calcSingleEvaluatorScore({ violations }) };
+}
