+++ packages/@n8n/ai-workflow-builder.ee/src/tools/test/validate-workflow.tool.test.ts
@@ -46,6 +46,7 @@ describe('validateWorkflow tool', () => {
 
 	const sampleValidationResult: ProgrammaticChecksResult = {
 		connections: [],
+		nodes: [],
 		trigger: [],
 		agentPrompt: [
 			{
