+++ packages/@n8n/ai-workflow-builder.ee/src/validation/checks/index.ts
@@ -1,8 +1,6 @@
 export { validateAgentPrompt } from './agent-prompt';
 export { validateConnections } from './connections';
 export { validateFromAi } from './from-ai';
+export { validateNodes } from './nodes';
 export { validateTools } from './tools';
-export {
-	validateTrigger,
-	type TriggerEvaluationResult,
-} from './trigger';
+export { validateTrigger, type TriggerEvaluationResult } from './trigger';
