+++ packages/@n8n/ai-workflow-builder.ee/src/validation/checks/trigger.ts
@@ -17,16 +17,8 @@ export function validateTrigger(
 ): ProgrammaticViolation[] {
 	const violations: ProgrammaticViolation[] = [];
 	const triggerNodes: string[] = [];
-	const nodeCountByType = new Map<string, number>();
 
 	if (!workflow.nodes || workflow.nodes.length === 0) {
-		violations.push({
-			name: 'workflow-has-no-nodes',
-			type: 'critical',
-			description: 'Workflow has no nodes',
-			pointsDeducted: 50,
-		});
-
 		return violations;
 	}
 
@@ -40,9 +32,6 @@ export function validateTrigger(
 		if (isTriggerNode(nodeType)) {
 			triggerNodes.push(node.name);
 		}
-
-		const currentCount = nodeCountByType.get(node.type) ?? 0;
-		nodeCountByType.set(node.type, currentCount + 1);
 	}
 
 	const hasTrigger = triggerNodes.length > 0;
@@ -54,23 +43,6 @@ export function validateTrigger(
 			description: 'Workflow must have at least one trigger node to start execution',
 			pointsDeducted: 50,
 		});
-	}
-
-	for (const [nodeTypeName, count] of nodeCountByType) {
-		const nodeType = nodeTypes.find((type) => type.name === nodeTypeName);
-
-		if (!nodeType?.maxNodes) {
-			continue;
-		}
-
-		if (count > nodeType.maxNodes) {
-			violations.push({
-				name: 'workflow-exceeds-max-nodes-limit',
-				type: 'critical',
-				description: `Workflow can only have ${nodeType.maxNodes} ${nodeType.displayName} node(s), but found ${count}`,
-				pointsDeducted: 50,
-			});
-		}
 	}
 
 	return violations;
